Move comments DAV plugins to comments app
