+++ apps/comments/lib/Dav/EntityTypeCollection.php
@@ -20,7 +20,7 @@
  *
  */
 
-namespace OCA\DAV\Comments;
+namespace OCA\Comments\Dav;
 
 use OCP\Comments\ICommentsManager;
 use OCP\ILogger;
@@ -38,7 +38,7 @@ use Sabre\DAV\Exception\NotFound;
  * Its children are instances of EntityCollection (representing a specific
  * object, for example the file by id).
  *
- * @package OCA\DAV\Comments
+ * @package OCA\Comments\Dav
  */
 class EntityTypeCollection extends RootCollection {
